Improvements to rt-scope, rt-class
A number of fixes rolled up into one commit. I'll submit separate PRs
for this if requested.

1. Allow trailing semicolon in rt-scope

Rationale: it is natural for many developers to place a trailing
semicolon after the last scope alias pair, just as they do after every
preceding pair in the list. Since it is possible to save them time (and
perhaps frustration) at no cost in terms of semantics and safety, we
should do so.

2. Merge the class and rt-class attributes

Previously an error was raised when a node had both the class and the
rt-class attribute defined. This leads to unnecessary verbose code which
taxes the reader. Splitting static class names from dynamic ones removes
this distraction, and keeps the semantics clean in the reader's mind.

On renaming .*Props to .*Attrs. I would have liked to avoid such global
renames on my first commit to somebody else's code base, but I needed to
add a constant for "className", and that seemed like a property on a
React object, as opposed to an attribute on an html element that only
later gets converted to a property. So I took the liberty to rename the
variables.

On tests. Having noticed how order-depenedent some of the code is (in
particular when it comes to dealing with rt-if/rt-repeat/rt-scope), I
decided it might be a good idea to be somewhat black-boxy in the tests,
and ensure that merging of class names works correctly in both
directions.

3. In rt-scope subsequent aliases should be able to reference preceding ones

Rationale: It is typical to want to alias multiple subtrees of a nested
structure in a single rt-scope statement, like so:

<div rt-scope="this.props.user as user; user.profile as profile;
user.friends as friends">

This was not possible because of the function parameter based
implementation of rt-scope. I tweaked it a bit to use var declaration
instead, while preserving the parameter-based passing for all child
scopes.

4. Only evaluate rt-scope after rt-if condition passed

Rationale: The typical use case for the combination of rt-if and
rt-scope is the "if the condition passes, let me alias the variables
that I'll need to build the inside of the block". Yet in the current
implementation if one assumes that the condition is true when writing
the expressions in rt-scope, one will be faced with errors, since the
scope is evaluated irrespective of the condition. This leads to clumsy
and difficult-to-read verbosity. Ideal semantics can be illustrated by
the following example:

<div rt-if="user.profile" rt-scope="user.profile as profile; profile.photo as photo;
profile.friends as friends;">

This is a breaking change. The use case that will be broken is:

<div rt-if="user.isLoggedIn" rt-scope="data.user as user">

that is, one in which the condition relies on a scope variable.

There's no ideal solution in this case — either the rt-if or the
rt-scope has to be evaluated first. Given that rt-scope is mostly used
for aliasing variables to be used in the child elements, which will only
render if the rt-if condition evaluates to true, it seems reasonable to
optimize for the more reader/writer-friendly rt-scope.

5. rt-scope should be evaluated in context of each rt-repeat's iteration

Rationale: similar to the change regarding the precedence of rt-if and
rt-scope, if we agree that rt-scope is mainly used to bind variables for
the use in the children, then it makes sense to have it per-iteration,
rather than as the top level over the whole rt-repeat.

This change is breaking the use case where rt-scope is used to define a
variable that is then referenced in rt-repeat. The idea is that such a
use case is rare, and the one described above is far more common.