Allow trailing semicolon in rt-scope
Rationale: it is natural for many developers to place a trailing
semicolon after the last scope alias pair, just as they do after every
preceding pair in the list. Since it is possible to save them time (and
perhaps frustration) at no cost in terms of semantics and safety, we
should do so.

On tests. Wasn't sure if this should be broken out into a separate test,
or bundled together with the existing one.