Only evaluate rt-scope after rt-if condition passed
Rationale: The typical use case for the combination of rt-if and
rt-scope is the "if the condition passes, let me alias the variables
that I'll need to build the inside of the block". Yet in the current
implementation if one assumes that the condition is true when writing
the expressions in rt-scope, one will be faced with errors, since the
scope is evaluated irrespective of the condition. This leads to clumsy
and difficult-to-read verbosity. Ideal semantics can be illustrated by
the following example:

<div rt-if="user.profile" rt-scope="user.profile as profile; profile.photo as photo;
profile.friends as friends;">

This is a breaking change. The use case that will be broken is:

<div rt-if="user.isLoggedIn" rt-scope="data.user as user">

that is, one in which the condition relies on a scope variable.

There's no ideal solution in this case — either the rt-if or the
rt-scope has to be evaluated first. Given that rt-scope is mostly used
for aliasing variables to be used in the child elements, which will only
render if the rt-if condition evaluates to true, it seems reasonable to
optimize for the more reader/writer-friendly rt-scope.